抢红包
Summary: 抢红包

Test Plan: local test

Reviewers: chenqixiang, yingming.zhao, lin.huang, lihaifeng

Reviewed By: chenqixiang

Differential Revision: https://pha.static.kunlun.com/D16088

+++ im/src/main/java/com/opay/im/model/LuckyMoneyModel.java
@@ -49,7 +49,7 @@ public class LuckyMoneyModel {
     * 目标类型:0个人 1群
     */
     @ApiModelProperty(value="目标类型:0个人 1群")
-    private Byte targetType;
+    private Integer targetType;
 
     /**
     * 红包说明Best Wishes!
